too many udpates to say

+++ HPD-AI-Framework/dotnet/shared/src/Helium/test/Helium.Primitives.Tests/MultisetTests.cs
@@ -9,7 +9,7 @@ public class MultisetTests
     [Fact]
     public void EmptyMultiset()
     {
-        var m = Multiset<int>.Empty;
+        var m = Multiset<Integer>.Empty;
         Assert.True(m.IsEmpty);
         Assert.Equal(0, m.Card);
     }
@@ -17,7 +17,7 @@ public class MultisetTests
     [Fact]
     public void FromElementsWithDuplicates()
     {
-        var m = Multiset<int>.FromElements([1, 1, 2]);
+        var m = Multiset<Integer>.FromElements([1, 1, 2]);
         Assert.Equal(2, m.Count(1));
         Assert.Equal(1, m.Count(2));
         Assert.Equal(3, m.Card);
@@ -28,14 +28,14 @@ public class MultisetTests
     [Fact]
     public void CountAbsentElementReturnsZero()
     {
-        var m = Multiset<int>.FromElements([1, 2]);
+        var m = Multiset<Integer>.FromElements([1, 2]);
         Assert.Equal(0, m.Count(99));
     }
 
     [Fact]
     public void CardIsSumOfCounts()
     {
-        var m = Multiset<int>.FromElements([1, 1, 1, 2, 2]);
+        var m = Multiset<Integer>.FromElements([1, 1, 1, 2, 2]);
         Assert.Equal(5, m.Card);
     }
 
@@ -44,7 +44,7 @@ public class MultisetTests
     [Fact]
     public void AddIncrementsCount()
     {
-        var m = Multiset<int>.FromElements([1, 2]);
+        var m = Multiset<Integer>.FromElements([1, 2]);
         var m2 = m.Add(1);
         Assert.Equal(2, m2.Count(1));
         Assert.Equal(1, m2.Count(2));
@@ -53,7 +53,7 @@ public class MultisetTests
     [Fact]
     public void RemoveDecrementsCount()
     {
-        var m = Multiset<int>.FromElements([1, 1, 2]);
+        var m = Multiset<Integer>.FromElements([1, 1, 2]);
         var m2 = m.Remove(1);
         Assert.Equal(1, m2.Count(1));
     }
@@ -61,7 +61,7 @@ public class MultisetTests
     [Fact]
     public void RemoveLastOccurrenceRemovesElement()
     {
-        var m = Multiset<int>.FromElements([1, 2]);
+        var m = Multiset<Integer>.FromElements([1, 2]);
         var m2 = m.Remove(1);
         Assert.Equal(0, m2.Count(1));
     }
@@ -69,7 +69,7 @@ public class MultisetTests
     [Fact]
     public void RemoveAbsentElementNoChange()
     {
-        var m = Multiset<int>.FromElements([1, 2]);
+        var m = Multiset<Integer>.FromElements([1, 2]);
         var m2 = m.Remove(99);
         Assert.Equal(m, m2);
     }
@@ -79,8 +79,8 @@ public class MultisetTests
     [Fact]
     public void UnionIsMaxOfCounts()
     {
-        var a = Multiset<int>.FromElements([1, 1, 2]);
-        var b = Multiset<int>.FromElements([1, 2, 2, 3]);
+        var a = Multiset<Integer>.FromElements([1, 1, 2]);
+        var b = Multiset<Integer>.FromElements([1, 2, 2, 3]);
         var u = a.Union(b);
         Assert.Equal(2, u.Count(1)); // max(2,1)
         Assert.Equal(2, u.Count(2)); // max(1,2)
@@ -90,8 +90,8 @@ public class MultisetTests
     [Fact]
     public void InterIsMinOfCounts()
     {
-        var a = Multiset<int>.FromElements([1, 1, 2]);
-        var b = Multiset<int>.FromElements([1, 2, 2, 3]);
+        var a = Multiset<Integer>.FromElements([1, 1, 2]);
+        var b = Multiset<Integer>.FromElements([1, 2, 2, 3]);
         var i = a.Inter(b);
         Assert.Equal(1, i.Count(1)); // min(2,1)
         Assert.Equal(1, i.Count(2)); // min(1,2)
@@ -101,8 +101,8 @@ public class MultisetTests
     [Fact]
     public void SumIsSumOfCounts()
     {
-        var a = Multiset<int>.FromElements([1, 1, 2]);
-        var b = Multiset<int>.FromElements([1, 2, 3]);
+        var a = Multiset<Integer>.FromElements([1, 1, 2]);
+        var b = Multiset<Integer>.FromElements([1, 2, 3]);
         var s = a.Sum(b);
         Assert.Equal(3, s.Count(1)); // 2+1
         Assert.Equal(2, s.Count(2)); // 1+1
@@ -115,16 +115,16 @@ public class MultisetTests
     [Fact]
     public void MapCollapsingElementsSumsCounts()
     {
-        var m = Multiset<int>.FromElements([1, 2, 3]);
+        var m = Multiset<Integer>.FromElements([1, 2, 3]);
         // Map all to the same value.
-        var mapped = m.Map(_ => 0);
+        var mapped = m.Map(_ => (Integer)0);
         Assert.Equal(3, mapped.Count(0));
     }
 
     [Fact]
     public void FilterKeepsSatisfyingElements()
     {
-        var m = Multiset<int>.FromElements([1, 2, 3, 4, 5]);
+        var m = Multiset<Integer>.FromElements([1, 2, 3, 4, 5]);
         var even = m.Filter(x => x % 2 == 0);
         Assert.Equal(1, even.Count(2));
         Assert.Equal(1, even.Count(4));
@@ -137,7 +137,7 @@ public class MultisetTests
     [Fact]
     public void ToFinsetDropsMultiplicity()
     {
-        var m = Multiset<int>.FromElements([1, 1, 2]);
+        var m = Multiset<Integer>.FromElements([1, 1, 2]);
         var fs = m.ToFinset();
         Assert.Equal(2, fs.Card);
         Assert.True(fs.Contains(1));
@@ -149,16 +149,16 @@ public class MultisetTests
     [Fact]
     public void EqualMultisets()
     {
-        var a = Multiset<int>.FromElements([1, 2, 1]);
-        var b = Multiset<int>.FromElements([1, 1, 2]);
+        var a = Multiset<Integer>.FromElements([1, 2, 1]);
+        var b = Multiset<Integer>.FromElements([1, 1, 2]);
         Assert.Equal(a, b);
     }
 
     [Fact]
     public void DifferentMultiplicitiesNotEqual()
     {
-        var a = Multiset<int>.FromElements([1, 2]);
-        var b = Multiset<int>.FromElements([1, 1, 2]);
+        var a = Multiset<Integer>.FromElements([1, 2]);
+        var b = Multiset<Integer>.FromElements([1, 1, 2]);
         Assert.NotEqual(a, b);
     }
 }
